+++ driver-core/CHANGELOG.rst
@@ -31,6 +31,8 @@ CHANGELOG
 - [improvement] Slow Query Logger (JAVA-646)
 - [improvement] Exclude some errors from measurements in LatencyAwarePolicy
   (JAVA-698)
+- [bug] Fix issue when executing a PreparedStatement from another cluster
+  (JAVA-641)
 
 Merged from 2.0.9_fixes branch: